requirements.txt
fastapi[all]
uvicorn
arrow
loguru
pandas
pandera[fastapi]
python-telegram-bot[http2]
motor